ui-tests- move tests to new suite
