+++ testing/specs/modal-dialog/content.duplicate.spec.js
@@ -3,12 +3,12 @@
  */
 const chai = require('chai');
 const assert = chai.assert;
-const webDriverHelper = require('../libs/WebDriverHelper');
-const appConstant = require('../libs/app_const');
-const studioUtils = require('../libs/studio.utils.js');
-const ContentBrowsePanel = require('../page_objects/browsepanel/content.browse.panel');
-const ContentDuplicateDialog = require('../page_objects/content.duplicate.dialog');
-const contentBuilder = require("../libs/content.builder");
+const webDriverHelper = require('../../libs/WebDriverHelper');
+const appConstant = require('../../libs/app_const');
+const studioUtils = require('../../libs/studio.utils.js');
+const ContentBrowsePanel = require('../../page_objects/browsepanel/content.browse.panel');
+const ContentDuplicateDialog = require('../../page_objects/content.duplicate.dialog');
+const contentBuilder = require("../../libs/content.builder");
 
 describe('content.duplicate.spec: Select and duplicate 2 folders', function () {
     this.timeout(appConstant.SUITE_TIMEOUT);
