+++ testing/specs/modal-dialog/delete.content.dialog.spec.js
@@ -3,13 +3,13 @@
  */
 const chai = require('chai');
 const assert = chai.assert;
-const webDriverHelper = require('../libs/WebDriverHelper');
-const appConstant = require('../libs/app_const');
-const ContentBrowsePanel = require('../page_objects/browsepanel/content.browse.panel');
-const studioUtils = require('../libs/studio.utils.js');
-const contentBuilder = require("../libs/content.builder");
-const DeleteContentDialog = require('../page_objects/delete.content.dialog');
-const PublishContentDialog = require('../page_objects/content.publish.dialog');
+const webDriverHelper = require('../../libs/WebDriverHelper');
+const appConstant = require('../../libs/app_const');
+const ContentBrowsePanel = require('../../page_objects/browsepanel/content.browse.panel');
+const studioUtils = require('../../libs/studio.utils.js');
+const contentBuilder = require("../../libs/content.builder");
+const DeleteContentDialog = require('../../page_objects/delete.content.dialog');
+const PublishContentDialog = require('../../page_objects/content.publish.dialog');
 
 describe('delete.content.dialog.spec:  tests for Delete Content Dialog', function () {
     this.timeout(appConstant.SUITE_TIMEOUT);
