+++ testing/specs/modal-dialog/wizard.compare.versions.dialog.spec.js
@@ -3,15 +3,15 @@
  */
 const chai = require('chai');
 const assert = chai.assert;
-const webDriverHelper = require('../libs/WebDriverHelper');
-const appConst = require('../libs/app_const');
-const contentBuilder = require("../libs/content.builder");
-const studioUtils = require('../libs/studio.utils.js');
-const ContentWizard = require('../page_objects/wizardpanel/content.wizard.panel');
-const WizardDetailsPanel = require('../page_objects/wizardpanel/details/wizard.details.panel');
-const WizardVersionsWidget = require('../page_objects/wizardpanel/details/wizard.versions.widget');
-const CompareContentVersionsDialog = require('../page_objects/compare.content.versions.dialog');
-const SettingsStepForm = require('../page_objects/wizardpanel/settings.wizard.step.form');
+const webDriverHelper = require('../../libs/WebDriverHelper');
+const appConst = require('../../libs/app_const');
+const contentBuilder = require("../../libs/content.builder");
+const studioUtils = require('../../libs/studio.utils.js');
+const ContentWizard = require('../../page_objects/wizardpanel/content.wizard.panel');
+const WizardDetailsPanel = require('../../page_objects/wizardpanel/details/wizard.details.panel');
+const WizardVersionsWidget = require('../../page_objects/wizardpanel/details/wizard.versions.widget');
+const CompareContentVersionsDialog = require('../../page_objects/compare.content.versions.dialog');
+const SettingsStepForm = require('../../page_objects/wizardpanel/settings.wizard.step.form');
 
 describe('wizard.compare.versions.dialog - open the dialog and verify elements', function () {
     this.timeout(appConst.SUITE_TIMEOUT);
